backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
python-dotenv==1.0.1
sqlalchemy==2.0.29
asyncpg==0.29.0  # For PostgreSQL async support
pydantic==2.7.0
pydantic-settings==2.2.1
aiofiles==23.2.1  # For file uploads
httpx==0.27.0     # Async HTTP client
spacy==3.8.2      # AI/nlp logic for Buddy
scikit-learn==1.4.2  # For recommendation and classification
numpy==1.24.4
pandas==2.0.3  # For data manipulation
